Move slider models to widgets/sliders.py
