+++ bokeh/models/widgets/__init__.py
@@ -8,5 +8,6 @@ from .icons import *
 from .inputs import *
 from .markups import *
 from .panels import *
+from .sliders import *
 from .tables import *
 from .widget import *
